feat(sw): de2shell firmware — unified shell with expdemo and crypto
- main.c: multi-channel program dispatch, IR frequency switching
- crypto.c: AES/SHA/SM4 with Zk* hardware acceleration
- dashboard.c: real-time board status display
- expdemo entry via demo.c, removed standalone exp1/4/5/12 commands
- monitor.c: RISC-V register peek/poke + Zk* instruction demo
- ps2.c: keyboard scancode monitor with lock LED sync
- life.c: Conway's Game of Life with editor mode
- ntt.c: NTT accelerator driver (SW reference + HW MMIO)
- vga_hal: text terminal abstraction layer
- conway_ed.c: removed (merged into life.c)

+++ sw/app/de2shell/ps2.c
@@ -98,19 +98,19 @@ static void log_maybe_roll(void) {
              VGA_GRAY);
 }
 
-static void log_puts(const char *s, uint8_t color) {
+static void log_puts(const char *s, uint16_t color) {
     log_maybe_roll();
     vga_puts(s, color);
 }
 
-static void log_put_hex8(uint8_t val, uint8_t color) {
+static void log_put_hex8(uint8_t val, uint16_t color) {
     static const char hex[] = "0123456789ABCDEF";
     log_maybe_roll();
     vga_putc(hex[(val >> 4) & 0x0f], color);
     vga_putc(hex[val & 0x0f], color);
 }
 
-static void log_put_dec_u8(uint8_t val, uint8_t color) {
+static void log_put_dec_u8(uint8_t val, uint16_t color) {
     char buf[3];
     int pos = 0;
